requirements.txt
fastapi
uvicorn
PyPDF2
streamlit
langchain
langchain-google-genai
faiss-cpu
python-dotenv
requests
